PlotSpikes.py: fix camel case conversion
In python3, one can no longer modify a dictionary while it is being
iterated. So we make a copy of the dictionary to get the iterator and
then we make whatever changes we want to make to the primary dictionary.

I also suspect that the stochastic failing before was because the
dictionary was being modified while it was being iterated on.

Fixes https://github.com/NeuroML/pyNeuroML/issues/68