enh: allow pynml to process multiple files
This tweaks the command line argument parsing to allow us to take
multiple NML/LEMS files as input in the shared options. For all the
options that can be applied to these, we then iterate over the file
list.

A few options don't use the file list given at all, so we handle them
separately.