feat: allow jnml output to be returned
This tweaks a few functions to allow jnml output to be returned for
programmatic access. At the moment, even though lots of things use jnml,
we only return it for validation.

This is not a breaking/backwards incompatible change since it does not
change the function signature or return value by default.
It is therefore, a backwards compatible enhancement.

Fixes #126