fix: do not pass NeuroMLDocument for pp based parallel runs
The complete libNeuroML API is based on lxml.etree now, and this
unfortunately cannot be pickled. So we cannot pass the NeuroMLDocument
to the run_simulation function when using pp for parallel runs.

Instead, we pass None and let the run_simulation method read in the
template NeuroML file and create the NeuroMLDocument there as required.