refactor(pynml): move all utility functions to separate file
This is required to prevent circular imports. For example, to include
the swc exporter in pynml, we need to import from pyneuron.swc.., but
that already needs read_neuroml2_file, which if in pynml.py will result
in a circular import.

So, pynml is now just a thin wrapper for the cli.
Note that there's no change in API here because pynml imports * from
utils.