fix(morph-plot): provide axis to colorbar
to handle deprecation warning:

```
PlotMorphology.py:784:
MatplotlibDeprecationWarning: Unable to determine Axes to steal space
for Colorbar. Using gca(), but will raise in the future. Either provide
the *cax* argument to use as the Axes for the Colorbar, provide the *ax*
argument to steal space from it, or add *mappable* to an Axes.
fig.colorbar()
```