refactor(plotting): split out on-demand loading to separate function
The logic here is that if a user provides us a file to plot, we don't
need to load all the includes in the file and so on.  We can load only
the information we need to plot the morphology plots and disregard the
other information.

However, if a loaded model is given, we just use that---it's already in
memory, and there's no point of copying it to remove things, since
that'll just use even more memory.