fix(vispy): use proximal and distal radii for mesh
During refactoring, we'd used only the one radii, but that meant we only
got cylinders, not cones and other related shapes where the two radii
can be different.

Note that this means there may be more unique meshes, and so, the
precision may be reduced more by the heuristic. But, it will allow users
with more computing power to get more accurate meshes.