feat(annotations): add bqbiol.is
https://github.com/combine-org/combine-specifications/blob/main/specifications/qualifiers-1.1.md#is-1

```
is

The biological entity represented by the model element has identity with
the subject of the referenced resource (biological entity B). This
relation might be used to link a reaction to its exact counterpart in a
database, for instance.
```

Note that since `is` is a Python keyword, this is implemented as `Is`,
with the argument being `is_`. Hopefully we're not picky about case
sensitivity in the rdf annotation..