feat(vispy-3d-detailed): take advantage of lru_cache
We make the function arguments all hashable so that we can use the
lru_cache here also. For most networks, there will be one cell object
that is placed at different locations in the population by specifying an
offset. The mesh creator takes the offset into account, but other than
that, it is not used. So, we don't need to pass it to our cell specific
method. For all cells of the same type for the population, we only
process the first one, and the rest of the meshes are taken from the
cache. We then add the offset for each cell instance to the mesh data
before passing it to the mesh creator.

So, for 100 cells of the same type, we only process the one cell.
The other 99 are picked from the cache.

For the moment, I've set the lru size to 100. I don't think we'll come
across networks that use more than 100 cells anytime soon.

Note that this does not help for cases where the same cell has also been
rotated because for each of these cells, NeuroML requires a new
population. There, we must process each cell individually, for each
population.