fix(swc2nml): correct generation of multi-point soma segments
The first two points belong to the first segment in this case also, and
that needed to be correctly recorded. Otherwise, other segments that
have the second point as the parent were not referring to the first
segment as the parent segment---they were referring to the next segment
where the second point was the proximal, not the distal.