feat(vispy): cast `axes_pos` to tuple to ensure it's a hashable object
This is required because all the arguments of functions that use caching
must be hashable. Now, the user should see the required type for the
`axes_pos` argument is a tuple, but if they miss that and provide a
list, they'll get a `not hashable` error, which isn't the most clear.

So, we help the user out by casting a list into a tuple where possible.