feat(vispy): simplify mesh creation
Move the calculations for constant plot types to the cells, and make
soma thicker to make it easy to see.

The mesh creator is now much more specific and does not need to know
what type of plot it's working on. It simply takes a list and spits out
a mesh.