fix(file-list): correct recursive file listing
When there are models that follow a recursive structure. For example:

```
something.net.nml -> ./cells/cell.nml -> ./channels/channel.nml
```

When processing the cell.nml file, we must do it relative to its
location, not relative to the root file (something.net.nml) so that the
paths are correct in the final list of files included in the model.

We do this by making sure that when recursing through the list, we pass
the location of the "new" rootfile, and when we append the processed
file to the file list, we include its "rootdir", which would be relevant
to the "main" rootfile.