fix(get-file-list): correct logic to handle different scenarios
Scenario 1:

Function is called from the directory where the "rootfile" lives.
This is simple, as all the files will have includes relative to this
directory. However, we do need to handle recursion because models and
includes can be split into any numbers of folders/levels.

Scenario 2:

Function is not called from the directory where the "rootfile" lives.
The directory is specified as the "rootdir". Here, we need to append the
rootdir to all the includes also so that they can be correctly found.

This case is not suggested, especially for archiving.